requirements.txt
streamlit
toml
requests
python-pptx
pymupdf
docx2txt
fitz
pdf2image